Refactor: 헤더 컴포넌트 및 oauth 파일 리팩토링
