--- src/components/header/Hamburger.tsx
@@ -1,12 +1,0 @@
-import HamburgerButtonLogo from './images/Hamburgerbutton.svg';
-import * as Ha from './styles/HeaderButtons.styles';
-
-function Hamburger() {
-  return (
-    <Ha.HeaderButtons>
-      <img src={HamburgerButtonLogo} alt="Hamburger Button" />
-    </Ha.HeaderButtons>
-  );
-}
-
-export default Hamburger;
