--- src/components/header/Logo.tsx
@@ -1,7 +1,0 @@
-import * as Lo from './styles/Logo.styles';
-
-function Logo() {
-  return <Lo.Logo>로고</Lo.Logo>;
-}
-
-export default Logo;
