+++ src/components/header/styles/Header.styles.tsx
@@ -14,6 +14,21 @@ export const NavBar = styled.div`
   background-color: white;
 `;
 
+export const Logo = styled.div`
+  cursor: pointer;
+  font-size: 24px;
+`;
+
+export const HeaderButtons = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 32px;
+  height: 32px;
+  margin-left: 10px;
+  cursor: pointer;
+`;
+
 export const LeftSection = styled.div`
   flex: 1;
 `;
