--- src/components/header/styles/HeaderButtons.styles.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-export const HeaderButtons = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  margin-left: 10px;
-  cursor: pointer;
-`;
