--- src/components/header/styles/Logo.styles.tsx
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const Logo = styled.div`
-  cursor: pointer;
-  font-size: 24px;
-`;
