--- src/components/kakaoauth/kako.d.tsx
@@ -1,8 +1,0 @@
-declare global {
-  interface Window {
-    Kakao: any;
-  }
-}
-
-export {};
-//카카오 속성 타입 정의 때문에 선언한 파일
